+++ src/dataview/global.d.ts
@@ -1,10 +1,14 @@
 declare module '*.icon.svg' {
     const icon: import('react').ComponentType<CustomIconComponentProps | import('react').SVGProps<SVGSVGElement>>
+    
+    // eslint-disable-next-line
     export default icon
 }
 
 declare module '*.svg' {
     const text: string
+    
+    // eslint-disable-next-line
     export default text
 }